Only initialize default values once
If multiple Parse are executed over a single application, all default values (set by Default en Envar) are reset to the default. In some case, it may be useful to only set the values on the first parse and keep these values between evaluation.

The new options does not change the default behavior, but allows the user to choose if he wants to keep values between each run. It is also possible to manually reset the evaluation to the default if desired.